Refactor directory and package
Package into separate packages
