+++ docs/README.md
@@ -1,4 +1,4 @@
-# Sentinel User Guide
+# sentinel.Sentinel User Guide
 
 // Update the title above to match the actual product name
