+++ src/main/java/sentinel/Sentinel.java
@@ -1,3 +1,14 @@
+package sentinel;
+
+import sentinel.command.ByeCommand;
+import sentinel.command.Command;
+import sentinel.exception.SentinelException;
+import sentinel.storage.FileLoader;
+import sentinel.storage.FileWriter;
+import sentinel.ui.Ui;
+import sentinel.utils.Parser;
+import sentinel.utils.SentinelList;
+
 public class Sentinel {
     private static SentinelList list;
     private final Ui ui;
@@ -25,7 +36,7 @@ public class Sentinel {
             } catch (IllegalArgumentException e) {
                 ui.showUnrecognisedCommand();
             } catch (SentinelException e){
-//                ui.showError(e);
+//                sentinel.ui.showError(e);
             }
             ui.showLine();
             new FileWriter(list).saveTasks();
