+++ src/main/java/sentinel/api/GeminiAPI.java
@@ -1,3 +1,5 @@
+package sentinel.api;
+
 import java.io.BufferedReader;
 import java.io.FileReader;
 import java.io.IOException;
