+++ src/main/java/sentinel/command/HelpCommand.java
@@ -1,0 +1,15 @@
+package sentinel.command;
+
+import sentinel.ui.Ui;
+import sentinel.utils.SentinelList;
+
+public class HelpCommand extends Command {
+    public HelpCommand(Ui ui, SentinelList list) {
+        super(ui, list);
+    }
+
+    @Override
+    public void execute(String input) {
+        ui.showHelp();
+    }
+}
