+++ src/main/java/sentinel/exception/DeadlineException.java
@@ -1,0 +1,7 @@
+package sentinel.exception;
+
+public class DeadlineException extends SentinelException {
+    public DeadlineException(String s) {
+        super(s);
+    }
+}
