+++ src/main/java/sentinel/exception/EventException.java
@@ -1,0 +1,7 @@
+package sentinel.exception;
+
+public class EventException extends SentinelException {
+    public EventException(String s) {
+        super(s);
+    }
+}
