+++ src/main/java/sentinel/exception/SentinelException.java
@@ -1,0 +1,10 @@
+package sentinel.exception;
+
+public abstract class SentinelException extends Exception{
+
+    public SentinelException(String s){
+        super(s);
+    }
+}
+
+
