+++ src/main/java/sentinel/storage/FileLoader.java
@@ -1,11 +1,16 @@
+package sentinel.storage;
+
 import java.io.BufferedReader;
 import java.io.FileReader;
 import java.io.IOException;
 import java.io.StringReader;
 import java.time.LocalDateTime;
-import java.util.ArrayList;
 import java.util.Properties;
-
+import sentinel.utils.SentinelList;
+import sentinel.task.Task;
+import sentinel.task.ToDo;
+import sentinel.task.Event;
+import sentinel.task.Deadline;
 public class FileLoader {
 
     public SentinelList loadTasks() {
@@ -36,7 +41,7 @@ public class FileLoader {
                 case "T" -> new ToDo(properties.getProperty("description"));
                 case "D" -> new Deadline(properties.getProperty("description"), LocalDateTime.parse(properties.getProperty("endDate")));
                 case "E" -> new Event(properties.getProperty("description"), LocalDateTime.parse(properties.getProperty("endDate")), LocalDateTime.parse(properties.getProperty("startDate")));
-                default -> throw new IllegalArgumentException("Unknown task type: " + taskType);
+                default -> throw new IllegalArgumentException("Unknown sentinel.task type: " + taskType);
             };
             if (properties.getProperty("isDone").equals("T")) task.setDone();
             tasks.add(task);
