+++ src/main/java/sentinel/task/Deadline.java
@@ -1,3 +1,5 @@
+package sentinel.task;
+
 import java.time.LocalDateTime;
 
 public class Deadline extends Task{
