+++ src/main/java/sentinel/task/Event.java
@@ -1,3 +1,5 @@
+package sentinel.task;
+
 import java.time.LocalDateTime;
 
 public class Event extends Task {
