+++ src/main/java/sentinel/task/ToDo.java
@@ -1,4 +1,6 @@
-public class ToDo extends Task {    
+package sentinel.task;
+
+public class ToDo extends Task {
     public ToDo(String description){
         super(description);
     }
