+++ src/main/java/sentinel/utils/Parser.java
@@ -1,3 +1,16 @@
+package sentinel.utils;
+
+import sentinel.api.GeminiAPI;
+import sentinel.exception.DeadlineException;
+import sentinel.exception.EventException;
+import sentinel.exception.SentinelException;
+import sentinel.task.Deadline;
+import sentinel.task.Event;
+import sentinel.task.Task;
+import sentinel.task.ToDo;
+import sentinel.ui.Ui;
+import sentinel.Sentinel;
+
 import java.time.LocalDateTime;
 import java.util.Arrays;
 
@@ -47,29 +60,29 @@ public class Parser {
 
         switch (commandType) {
             case todo -> {
-                // taskName is everything after the command
+                // taskName is everything after the sentinel.command
                 taskName = String.join(" ", Arrays.copyOfRange(stringArr, 1, stringArr.length)).trim();
             }
 
             case deadline -> {
-                // For 'deadline', taskName is everything after the command and before "/by"
+                // For 'deadline', taskName is everything after the sentinel.command and before "/by"
                 int byIndex = Arrays.asList(stringArr).indexOf("/by");
                 if (byIndex != -1) {
                     taskName = String.join(" ", Arrays.copyOfRange(stringArr, 1, byIndex)).trim();
                 } else {
                     ui.showDeadlineCommandGuidelines();
-                    throw new DeadlineException("Deadline Command");
+                    throw new DeadlineException("sentinel.task.Deadline sentinel.command.Command");
                 }
             }
 
             case event -> {
-                // For 'event', taskName is everything after the command and before "/from"
+                // For 'event', taskName is everything after the sentinel.command and before "/from"
                 int fromIndex = Arrays.asList(stringArr).indexOf("/from");
                 if (fromIndex != -1) {
                     taskName = String.join(" ", Arrays.copyOfRange(stringArr, 1, fromIndex)).trim();
                 } else {
                     ui.showEventCommandGuidelines();
-                    throw new EventException("Event Command");
+                    throw new EventException("sentinel.task.Event sentinel.command.Command");
                 }
             }
         }
