+++ text-ui-test/EXPECTED.TXT
@@ -20,13 +20,13 @@ ____________________________________________________________
 
 ____________________________________________________________
 
-Got it. I've added this task: CS2103T iP
+Got it. I've added this sentinel.task: CS2103T iP
         [T][ ] CS2103T iP
 ____________________________________________________________
 
 ____________________________________________________________
 
-Alright! I've marked this task as done:
+Alright! I've marked this sentinel.task as done:
         [X] CS2103T iP
 ____________________________________________________________
 
@@ -37,7 +37,7 @@ ____________________________________________________________
 
 ____________________________________________________________
 
-Got it. I've added this task: CS2101 Meeting
+Got it. I've added this sentinel.task: CS2101 Meeting
         [E][ ] CS2101 Meeting [from: 28 Aug 2024 2pm | to: 28 Aug 2024 5pm]
 ____________________________________________________________
 
@@ -48,7 +48,7 @@ ____________________________________________________________
 
 ____________________________________________________________
 
-Got it. I've added this task: YSC4231 Assignment
+Got it. I've added this sentinel.task: YSC4231 Assignment
         [D][ ] YSC4231 Assignment [by 2pm]
 ____________________________________________________________
 
@@ -62,7 +62,7 @@ ____________________________________________________________
 
 ____________________________________________________________
 
-Alright! I've marked this task as done:
+Alright! I've marked this sentinel.task as done:
         [X] YSC4231 Assignment
 ____________________________________________________________
 
@@ -73,7 +73,7 @@ ____________________________________________________________
 
 ____________________________________________________________
 
-Alright! I've marked this task as undone:
+Alright! I've marked this sentinel.task as undone:
         [ ] CS2103T iP
 ____________________________________________________________
 
@@ -89,7 +89,7 @@ ____________________________________________________________
 
 ____________________________________________________________
 
-I have deleted the following task:
+I have deleted the following sentinel.task:
         [T][ ] CS2103T iP
  You have 2 remaining tasks.
 ____________________________________________________________
