+++ text-ui-test/runtest.bat
@@ -15,7 +15,7 @@ IF ERRORLEVEL 1 (
 REM no error here, errorlevel == 0
 
 REM run the program, feed commands from input.txt file and redirect the output to the ACTUAL.TXT
-java -classpath ..\bin Sentinel < input.txt > ACTUAL.TXT
+java -classpath ..\bin sentinel.Sentinel < input.txt > ACTUAL.TXT
 
 REM compare the output to the expected output
 FC /W ACTUAL.TXT EXPECTED.TXT
